ingenieria_automatizacion/requirements.txt
# Análisis de datos y manipulación
pandas
numpy
openpyxl

# Visualización de datos
matplotlib
seaborn
plotly

# Machine Learning y Modelado estadístico
scipy
scikit-learn
statsmodels
xgboost
pingouin
watchdog

# Conectividad y Bases de Datos
sqlalchemy
pyodbc
psycopg2-binary
python-dotenv

# Motor para Jupyter Notebooks en VS Code
ipykernel